Allow selection of default profile using a file
For GUI versions of emacs, it's not possible to start them with a
`--with-profile` flag. In these cases, it's helpful to be able to
control which profile is loaded using something in the filesystem.
This change allows switching of the default profile using the contents
of a `~/.emacs-profile` file, which should contain the string of the
profile to be used (e.g. "spacemacs" or "doom" or "default").

If the file is not present, "default" will be used as before.