Fix bug in `.emacs-profile` handling
Having run `echo 'spacemacs' > ~/.emacs-profile`, the `.emacs` init script complained that it could not find a profile `spacemacs` in my set. I believe this is because the value was being turned into a symbol, e.g. `spacemacs` would not be found, but `"spacemacs"` could be. To fix this, I needed to turn the read value explicitly into a string.