Enable the loading a profile via CLI
Currently a profile must first be specified in `~/.emacs-profiles.el`
in order to be used. In most day-to-day usage this is adequate, but
on occasion it is nice to have a 'throw-away' profile. The rationale
for such a profile is trying someone else's config out for a day / hour
or some such, but it's not something you're committed to.

This change enables the following to work:
```bash
emacs --with-profile '((user-emacs-directory . "/tmp/throw-away-config"))'
```